refactor(lightbox): move Lightbox shell to features/lightbox
