+++ src/features/lightbox/index.tsx
@@ -1,0 +1,1 @@
+export {Lightbox} from './Lightbox'
